Added Console based reporting
