+++ TrainingPlan/TrainingPlan.BusinessModel/Topic.cs
@@ -9,5 +9,10 @@ namespace TrainingPlan.BusinessModel
     public class Topic
     {
         public string Name { get; set; }
+
+        public override string ToString()
+        {
+            return (Name);
+        }
     }
 }
