+++ TrainingPlan/TrainingPlan.BusinessModel/Trainee.cs
@@ -10,6 +10,11 @@
             }
         }
 
+        public override string ToString()
+        {
+            return (FirstName+" "+LastName);
+        }
+
     }
 
 }
