+++ TrainingPlan/TrainingPlan.BusinessModel/TrainingPlan.BusinessModel.csproj
@@ -40,12 +40,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ConsoleBasedReporting.cs" />
+    <Compile Include="ConsoleInput.cs" />
     <Compile Include="Coordinator.cs" />
+    <Compile Include="InMemoryDataStore.cs" />
     <Compile Include="Input.cs" />
     <Compile Include="IDataStore.cs" />
     <Compile Include="IReport.cs" />
     <Compile Include="Participant.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Report.cs" />
     <Compile Include="Topic.cs" />
     <Compile Include="TopicSchedule.cs" />
     <Compile Include="Trainee.cs" />
